+++ src/App.vue
@@ -4,7 +4,7 @@
 		<div class="main_chapter">
 			<Header />
 			<my-slider
-        numb="01"
+				numb="01"
 				title="race"
 				:arr="Object.values(this.race)"
 				:arr_keys="Object.keys(this.race)"
@@ -17,110 +17,25 @@
 			<div class="main_chapter_menu">
 				<RaceMenuSettings />
 			</div>
+
 			<transition name="btm-fade" mode="out-in">
-				<my-button v-if="pagesStore.shown_home" numb="02" title="class"></my-button>
+				<my-button
+					v-if="pagesStore.shown_home"
+					numb="02"
+					title="class"
+				></my-button>
 				<my-button-back v-else @click="showHome()"></my-button-back>
 			</transition>
 		</div>
 	</div>
 
 	<!-- Выпадающее меню -->
-	<div class="sidebar_wrap" :class="{ sidebar_wrap_open: pagesStore.setting_open }">
-		<!-- Превью -->
-		<my-selection-box :shown="main_page.shown.logo">
-			<Description />
-		</my-selection-box>
-		<!-- Превью -->
-
-		<!-- Смена языка -->
-		<my-selection-box :shown="main_page.shown.lang">
-      <LangSetting />
-		</my-selection-box>
-		<!-- Смена языка -->
-
-		<!-- Уровень -->
-		<my-selection-box :shown="main_page.shown.lvl">
-			<div class="flex_options">
-				<MyRange v-model.number="MY.level" lvl />
-				<MyRangeSize lvl />
-			</div>
-		</my-selection-box>
-		<!-- Уровень -->
-
-		<!-- Этнос-->
-		<my-selection-box :shown="race_page.shown.ethnos">
-			<EthnosChoice />
-		</my-selection-box>
-		<!-- Этнос -->
-
-		<!-- Гендр -->
-		<my-selection-box :shown="race_page.shown.gender">
-      <GenderSetting />
-		</my-selection-box>
-		<!-- Гендр -->
-
-		<!-- Цвет кожи -->
-		<my-selection-box :shown="race_page.shown.skin_color">
-			<my-color-select body_part="skin" />
-		</my-selection-box>
-		<!-- Цвет кожи -->
-
-		<!-- Цвет глаз -->
-		<my-selection-box :shown="race_page.shown.eyes_color">
-			<my-color-select body_part="eyes" />
-		</my-selection-box>
-		<!-- Цвет глаз -->
-
-		<!-- Цвет волос -->
-		<my-selection-box :shown="race_page.shown.hair_color">
-			<my-color-select body_part="hair" />
-		</my-selection-box>
-		<!-- Цвет волос -->
-
-		<!-- Возраст -->
-		<my-selection-box :shown="race_page.shown.age">
-			<div class="flex_options">
-				<MyRange v-model.number="MY.age" age />
-				<MyRangeSize age />
-			</div>
-		</my-selection-box>
-		<!-- Возраст -->
-
-		<!-- Рост -->
-		<my-selection-box :shown="race_page.shown.height">
-			<div class="flex_options">
-				<MyRange v-model.number="MY.height" height />
-				<mySizeGrowth />
-			</div>
-		</my-selection-box>
-		<!-- Рост -->
-
-		<!-- Вес -->
-		<my-selection-box :shown="race_page.shown.weight">
-			<div class="flex_options">
-				<MyRange v-model.number="MY.weight" weight />
-				<MyRangeSize weight />
-			</div>
-		</my-selection-box>
-		<!-- Вес -->
-
-		<!-- Характеристики -->
-		<my-selection-box :shown="race_page.shown.stats">
-			<RaceCustomStats />
-		</my-selection-box>
-		<!-- Характеристики -->
-
-		<!-- Навыки -->
-		<my-selection-box :shown="race_page.shown.skills">
-      <RaceCustomSkills />
-		</my-selection-box>
-		<!-- Навыки -->
-
-		<!-- Языки -->
-		<my-selection-box :shown="race_page.shown.languages">
-      <RaceCustomLanguages />
-		</my-selection-box>
-		<!-- Языки -->
+	<div
+		class="sidebar_wrap"
+		:class="{ sidebar_wrap_open: pagesStore.setting_open }"
+	>
+		<HeaderSettings />
+		<RaceSettings />
 	</div>
 
 	<div class="stripe"></div>
@@ -137,11 +52,11 @@
 				height: Char_Hight_Back,
 			}"
 		>
+			<WelcomeBanner />
+
 			<RaceBody body_part="skin" />
 			<RaceBody body_part="eyes" />
 			<RaceBody body_part="hair" />
-
-			<WelcomeBanner />
 
 			<transition name="slide-fade">
 				<mySizeGrowth v-if="hide_Ruler" division zero skale_top />
@@ -156,153 +71,26 @@
 		class="sidebar_right"
 		:class="{ sidebar_right_close: !pagesStore.shown_home }"
 	>
-		<!-- stats -->
-		<my-wrapper hr>
-			<my-attribute
-				v-for="name in stats_Keys"
-				:key="name"
-				:title="name"
-				:type="`${name}_base`"
-				plus
-				:numb="stats_Race_Page_Numb(name)"
-				:icon="name"
-			>
-			</my-attribute>
-		</my-wrapper>
-		<!-- stats -->
-
-		<!-- attributes_race -->
-		<my-wrapper hr v-if="skills_All_RE.length !== 0">
-			<my-attribute
-				v-for="name in skills_All_RE"
-				:key="name"
-				:title="name"
-				plus
-				:numb="Skill_Mastery"
-				:icon="MY.skills[name].mod"
-			></my-attribute>
-		</my-wrapper>
-		<!-- attributes_race -->
-
-		<!-- qualities -->
-		<my-wrapper hr>
-			<my-attribute
-				v-for="(val, name) in MY.qualities"
-				:key="name"
-				:title="name"
-				:numb="getParNumb('qualities', name)"
-				feet
-				:icon="name"
-			></my-attribute>
-			<my-attribute
-				v-if="MY.ethnos.hp_bonus"
-				title="hp_bonus"
-				:numb="hp_bonus"
-				plus
-				icon="hp_bonus"
-			></my-attribute>
-		</my-wrapper>
-		<!-- qualities -->
-
-		<!-- proficiencies -->
-		<my-wrapper gap_8 hr>
-			<my-inventory
-				v-for="(val, name) in MY.proficiencies"
-				:key="name"
-				:title="name"
-				:item="getProf_REX(name)"
-			>
-			</my-inventory>
-		</my-wrapper>
-		<!-- proficiencies -->
-
-		<!-- fines -->
-		<my-wrapper v-if="MY.race.fines || MY.ethnos.fines" gap_8 hr>
-			<my-fines
-				v-for="item in MY.race.fines"
-				:key="item"
-				:icon="item.type"
-				:title="item.keyword"
-				:details="item.details"
-			></my-fines>
-
-			<my-fines
-				v-for="item in MY.ethnos.fines"
-				:key="item"
-				:icon="item.type"
-				:title="item.keyword"
-				:details="item.details"
-			></my-fines>
-		</my-wrapper>
-		<!-- fines -->
-
-		<!-- spells -->
-		<my-wrapper v-if="showSpells" gap_26 hr>
-			<my-spell-text
-				v-for="item in MY.race.spells"
-				:key="item"
-				:lvl="item.level"
-				:spell="item.spell"
-			>
-			</my-spell-text>
-			<my-spell-text
-				v-for="item in MY.ethnos.spells"
-				:key="item"
-				:lvl="item.level"
-				:spell="item.spell"
-			>
-			</my-spell-text>
-		</my-wrapper>
-		<!-- spells -->
-
-		<!-- text -->
-		<div class="story int-400">
-			<div v-html="t(MY.race.details)"></div>
-			<my-card-text
-				v-if="MY.ethnos.name !== 'common'"
-				:title="MY.ethnos.name"
-				:text="MY.ethnos.details"
-				:rare="MY.ethnos.rare"
-			>
-			</my-card-text>
-		</div>
+		<RaceParameters />
 	</div>
 	<!-- </transition> -->
 	<!-- sidebar_right -->
 </template>
 
 <script>
-import dic from "@/assets/catalog/texts/dic.js";
-
-
 import race from "@/assets/catalog/base_data/step1_races.js";
 import clas from "@/assets/catalog/base_data/step2_classes.js";
-// import past from "@/assets/catalog/base_data/step3_backstories.js";
-// import languages from "@/assets/catalog/base_data/list_languages.js";
-// import placeholder from "@/assets/catalog/base_data/_placeholder.js";
 
-import EthnosChoice from "@/components/pinia/race_page/settings/EthnosChoice.vue";
-
-// import AgeWeight from "@/components/AgeWeight.vue";
-
-
-// PINIA
-import Header from "@/components/pinia/Header.vue";
-import Description from "@/components/pinia/Description.vue";
 import WelcomeBanner from "@/components/pinia/WelcomeBanner.vue";
 
-import LangSetting from "@/components/pinia/LangSetting.vue";
+import Header from "@/components/pinia/Header.vue";
+import HeaderSettings from "@/components/pinia/HeaderSettings.vue";
 
+// RACE_PAGE
 import RaceMenuSettings from "@/components/pinia/race_page/settings/RaceMenuSettings.vue";
-
-import RaceCustomStats from "@/components/pinia/race_page/settings/RaceCustomStats.vue";
-import RaceCustomSkills from "@/components/pinia/race_page/settings/RaceCustomSkills.vue";
-import RaceCustomLanguages from "@/components/pinia/race_page/settings/RaceCustomLanguages.vue";
-
-import GenderSetting from "@/components/pinia/race_page/settings/GenderSetting.vue";
-// PINIA
-
-
+import RaceSettings from "@/components/pinia/race_page/settings/RaceSettings.vue";
+import RaceParameters from "@/components/pinia/race_page/RaceParameters.vue";
+// RACE_PAGE
 
 import { watch } from "vue";
 import { mapState } from "pinia";
@@ -311,82 +99,64 @@ import { useMYStore } from "@/stores/MY/MYStore";
 import { useColorStore } from "@/stores/modules/ColorStore";
 export default {
 	setup() {
-    // PINIA
-    const pagesStore = usePagesStore();
-    const { main_page, race_page } = usePagesStore();
-    const { showHome, closeEthnos, closeColor, closePar, showRaceScroll } = usePagesStore();
-    const { MY } = useMYStore();
-    const { getEthnos } = useMYStore();
-    // PINIA
+		// PINIA
+		const pagesStore = usePagesStore();
+		const { main_page, race_page } = usePagesStore();
+		const { showHome, closeEthnos, closeColor, closePar, showRaceScroll } =
+			usePagesStore();
+		const { MY } = useMYStore();
+		const { getEthnos } = useMYStore();
+		// PINIA
 
-		watch(() => MY.race, () => {
-			console.log("Заметка обновилась!");
-			// console.log('closeEthnos:', closeEthnos)
-			getEthnos();
-			closeEthnos();
-			closeColor("skin");
-			closeColor("eyes");
-			closeColor("hair");
-			closePar("stats");
-			closePar("skills");
-			closePar("languages");
-		});
+		watch(
+			() => MY.race,
+			() => {
+				console.log("Заметка обновилась!");
+				// console.log('closeEthnos:', closeEthnos)
+				getEthnos();
+				closeEthnos();
+				closeColor("skin");
+				closeColor("eyes");
+				closeColor("hair");
+				closePar("stats");
+				closePar("skills");
+				closePar("languages");
+			}
+		);
 
-		return { 
-      showHome, 
-      showRaceScroll, 
+		return {
+			showHome,
+			showRaceScroll,
 
-      // PINIA
-      pagesStore, 
-      main_page,
-      race_page,
-      MY 
-    };
+			// PINIA
+			pagesStore,
+			main_page,
+			race_page,
+			MY,
+		};
 	},
 	name: "App",
 	components: {
-		// GenderChoiceStore,
-		EthnosChoice,
-		// AgeWeight,
-    
-		// ГОТОВ
-		WelcomeBanner, //^ DONE
-		Header, //^ DONE
-		Description, //^ DONE
-    
-    LangSetting, //^ DONE
-    
-    RaceCustomStats, //^ DONE
-    RaceCustomSkills, //TODO: Перенести Skill_Mastery
-    RaceCustomLanguages, //^ DONE
-    
-    GenderSetting, //^ DONE
-		// ГОТОВ
+		WelcomeBanner,
 
-		// НА ОБРАБОТКЕ
+		Header,
+		HeaderSettings,
+		// RACE_PAGE
 		RaceMenuSettings, //TODO: Сылки на рост, вес, возраст
-		// НА ОБРАБОТКЕ
+		RaceSettings,
+		RaceParameters,
+		// RACE_PAGE
+
 	},
 
 	data() {
 		return {
-			dic: dic,
 			race: race,
 			clas: clas,
-			// past: past,
-			// languages: languages,
-			// placeholder: placeholder,
-
 		};
 	},
 
 	created() {
-		// this.MY.race = Object.values(this.race)[0];
-		// this.MY.ethnos = Object.values(
-		// 	this.MY.race.settings.ethnos
-		// )[0];
-		// this.MY = this.default_MY;
-
 		this.MY.class = Object.values(clas)[1];
 		this.MY.mastery = this.Mastery;
 
@@ -394,24 +164,20 @@ export default {
 	},
 
 	computed: {
-    // ...mapState(useMYStore, ["MY"]),
-    ...mapState(useColorStore, ["color_Char_Сommon"]),
-    ...mapState(useMYStore, [
-      "stats_Keys",
-      "languages_Custom_Arr_RE",
-      "stats_Race_Page_Numb",
-      "skills_All_RE"
-    ]),
+		// ...mapState(useMYStore, ["MY"]),
+		...mapState(useColorStore, ["color_Char_Сommon"]),
+		...mapState(useMYStore, [
+			"stats_Keys",
+			"languages_Custom_Arr_RE",
+			"stats_Race_Page_Numb",
+			"skills_All_RE",
+		]),
 
 		Mastery() {
 			return Math.ceil(this.MY.level / 4);
 		},
 
-		Skill_Mastery() {
-			return 1 + this.MY.mastery;
-		},
-
-    hide_Ruler() {
+		hide_Ruler() {
 			return this.pagesStore.shown_home || this.race_page.shown.height;
 		},
 
@@ -446,19 +212,8 @@ export default {
 			}
 		},
 
-		hp_bonus() {
-			let increm_1 = this.MY.ethnos.hp_bonus[0];
-			let increm_2 = this.MY.ethnos.hp_bonus[1];
-			let level = Math.ceil(this.MY.level / increm_1);
-			return level * increm_2;
-		},
-
 		race_Settings() {
 			return this.MY.race.settings;
-		},
-
-		ethnos_Settings() {
-			return this.MY.ethnos;
 		},
 
 		Char_Hight_Back() {
@@ -475,28 +230,11 @@ export default {
 			} else {
 				kof = 3.1;
 			}
-			if (
-				this.race_page.shown.eyes_color ||
-				this.race_page.shown.hair_color
-			) {
+			if (this.race_page.shown.eyes_color || this.race_page.shown.hair_color) {
 				return `calc((100% / 210 * ${mein_height})*${kof})`;
 			} else {
 				return `100%`;
 			}
-		},
-
-		All_Ethnos_Obj() {
-			return this.MY.race.settings.ethnos;
-		},
-
-
-		showSpells() {
-			let race = this.MY.race.spells;
-			let ethnos = this.MY.ethnos.spells;
-			let lvl = this.MY.level;
-			let race_lvl = ((race || {})[0] || {}).level <= lvl;
-			let ethnos_lvl = ((ethnos || {})[0] || {}).level <= lvl;
-			return race || (ethnos && race_lvl) || ethnos_lvl;
 		},
 	},
 	watch: {
@@ -516,66 +254,6 @@ export default {
 			this.MY.weight = this.Get_Weight;
 			this.MY.age = this.Get_Age;
 		},
-
-		// getProficienciesItem(name) {
-		// 	let arr = [];
-		// 	for (let i in this.MY.race.proficiencies[name]) {
-		// 		arr.push(this.MY.race.proficiencies[name][i].name);
-		// 	}
-		// 	return arr;
-		// },
-
-		// getProficienciesEthnosItem(obj, name) {
-		// 	let arr = [];
-		// 	for (let i in this.All_Ethnos_Obj[obj].proficiencies[name]) {
-		// 		arr.push(this.All_Ethnos_Obj[obj].proficiencies[name][i].name);
-		// 	}
-		// 	return arr;
-		// },
-
-		getProf_Item(obj, kay) {
-			let arr = [];
-			for (let i in obj) {
-				arr.push(obj[i][kay]);
-			}
-			return arr;
-		},
-
-		getProfArr(obj, kay) {
-			let arr = [];
-			if (obj?.[kay]) {
-        arr = obj[kay].map(x => x.name)
-			}
-			return arr;
-		},
-
-		getProf_REX(kay) {
-			let i = this.getProfArr(this.MY.race.proficiencies, kay);
-			let j = this.getProfArr(this.MY.ethnos.proficiencies, kay);
-      let k = []
-      if (kay === "languages") {
-			k = this[`${kay}_Custom_Arr_RE`];
-    }
-			// let k = this.getProfArr(this.race_page.extra, kay);
-			return i.concat(j).concat(k);
-		},
-
-		getParNumb(par_1, par_2) {
-			let i = 0;
-			let j = 0;
-			if (((this.MY.race || {})[par_1] || {})[par_2]) {
-				i = this.MY.race[par_1][par_2];
-			}
-			if (((this.MY.ethnos || {})[par_1] || {})[par_2]) {
-				j = this.MY.ethnos[par_1][par_2];
-			}
-			return i + j;
-		},
-
-		getMannaNumb(arr, name) {
-			let index = arr.findIndex(el => el[name]);
-			return index;
-		},
 	},
 };
 </script>
@@ -589,30 +267,16 @@ a {
 	color: #fff;
 }
 
-.green {
-	color: green;
-}
-
-.skill_marg {
-	margin-top: 26px;
-}
-
 #app {
-	/*font-family: 'JetBrains Mono', sans-serif;*/
 	-webkit-font-smoothing: antialiased;
 	-moz-osx-font-smoothing: grayscale;
 	height: 100%;
 	display: flex;
 	-webkit-touch-callout: none;
-	/* iOS Safari */
 	-webkit-user-select: none;
-	/* Chrome/Safari/Opera */
 	-khtml-user-select: none;
-	/* Konqueror */
 	-moz-user-select: none;
-	/* Firefox */
 	-ms-user-select: none;
-	/* Internet Explorer/Edge */
 	user-select: none;
 	overflow: hidden;
 }
@@ -649,18 +313,11 @@ a {
 }
 
 .sidebar_left {
-	/* height: 100%; */
-	/*display: flex;*/
 	padding-top: 32px;
 	width: 320px;
 	background-color: #0e1518;
-	/*min-height: 100%;*/
 	display: flex;
 	flex-direction: column;
-	/*justify-content: space-between;*/
-	/*height: 100%;*/
-	/*z-index: 10;*/
-	/* outline: 2px solid rgba(255, 255, 255, 0.1); */
 }
 
 .main_menu_wrap {
@@ -677,31 +334,14 @@ a {
 	width: 0;
 }
 
-.chapter {
-	/* height: 100%; */
-	/*display: flex;*/
-	/*flex-direction: column;*/
-	/*justify-content: space-between;*/
-	/* flex: 1 1 auto; */
-}
-
 .main_chapter {
 	padding: 0 32px 0 32px;
 }
 
 .main_chapter_menu {
-	/* height: 100%; */
-	/*width: 320px;*/
 	padding: 40px 32px 32px 32px;
 	overflow-y: scroll;
 	max-height: 100%;
-
-	/*padding-left: 32px;*/
-	/*padding-top: 32px;*/
-	/*max-height: 100%;*/
-	/*flex: 1 1 auto;*/
-	/*overflow-y: scroll;*/
-	/*max-height: 100%;*/
 	scrollbar-width: none;
 }
 
@@ -710,14 +350,9 @@ a {
 }
 
 .selection_menu_wrap {
-	/*overflow-y: scroll;*/
 	display: flex;
 	flex-direction: column;
 	gap: 34px;
-	/*overflow-y: scroll;*/
-	/*height: 100px;*/
-	/*overflow: hidden;*/
-	/*max-height: 100%;*/
 }
 
 .selection_menu_wrap::-webkit-scrollbar {
@@ -725,38 +360,11 @@ a {
 }
 
 .selection_menu {
-	/* height: 100%; */
 	width: 256px;
 	display: flex;
 	flex-direction: column;
 	gap: 8px;
-	/*margin-bottom: 34px;*/
 }
-
-.ethnos_attributes {
-	color: rgba(255, 255, 255, 0.2);
-	display: flex;
-	flex-direction: column;
-	gap: 26px;
-	margin: 0 0 26px 16px;
-}
-
-.selection_card_active {
-	border: 2px solid #ffffff;
-	padding: 14px !important;
-}
-
-/* .mode-fade-enter-active,
-.mode-fade-leave-active {
-  transition: opacity .3s ease;
-}
-
-.mode-fade-enter-from,
-.mode-fade-leave-to {
-  transform: translateX(20px);
-  opacity: 0;
-
-} */
 
 .btm-fade-enter-active {
 	transition: all 0.2s ease-in-out;
@@ -765,10 +373,6 @@ a {
 .btm-fade-leave-active {
 	transition: all 0.2s ease-in-out;
 }
-
-/* .mode-fade-leave-active {
-  transition: all .3s cubic-bezier(1.0, 0.5, 0.8, 1.0);
-} */
 
 .btm-fade-enter-from,
 .btm-fade-leave-to {
@@ -784,7 +388,6 @@ a {
 	display: flex;
 	background-color: #0e1518;
 	z-index: 2;
-	/* position: relative; */
 	overflow: hidden;
 }
 
@@ -796,39 +399,8 @@ a {
 	transition-timing-function: ease-in-out;
 }
 
-/* .character img {
-  position: absolute;
-  bottom: 0;
-  right: 50%;
-  left: 50%;
-  -webkit-transform: translate(-50%, 0%);
-  -ms-transform: translate(-50%, 0%);
-  transform: translate(-50%, 0%);
-  transition-duration: 0.8s;
-  transition-timing-function: ease-in-out;
-} */
-
-/* .character svg {
-  position: absolute;
-  bottom: 0;
-  right: 50%;
-  left: 50%;
-  -webkit-transform: translate(-50%, 0%);
-  -ms-transform: translate(-50%, 0%);
-  transform: translate(-50%, 0%);
-  transition-property: all, fill;
-  transition-duration: 0.8s, 0.1s;
-  transition-timing-function: ease-in-out;
-} */
-
-/* .active_eyes .character img {
-  top: 0;
-  bottom: auto;
-} */
-
 .active_eyes {
 	align-self: flex-start;
-	/* height: 200%; */
 	transition-property: all;
 	transition-duration: 0.8s;
 	transition-timing-function: ease-in-out;
@@ -853,32 +425,10 @@ a {
 
 .sidebar_wrap_open {
 	width: 426px;
-	/* padding: 32px 32px 32px 0; */
 	transition: all 0.4s ease-in-out;
 }
 
-/* .sidebar_selection {
-  height: 100%;
-  position: absolute;
-  left: -426px;
-  transition: all 0.4s ease-in-out;
-  opacity: 0;
-}
-
-.sidebar_selection::-webkit-scrollbar {
-  width: 0;
-}
-
-.sidebar_selection_open {
-  opacity: 1;
-  left: 0px;
-  transition: all 1s ease-in-out;
-} */
-
-/* ---------------------sidebar_right----------------------*/
-
 .sidebar_right {
-	/* width: 426px; */
 	min-width: 426px;
 	max-width: 426px;
 	padding: 32px;
@@ -931,49 +481,6 @@ a {
 	opacity: 0;
 }
 
-/* -------------slider-grwwth------------------- */
-
-.slider_growth_back {
-	width: 344px;
-	height: calc(100% / 700 * 396);
-	background: rgba(255, 255, 255, 0.06);
-	border-radius: 12px;
-	display: flex;
-	align-items: flex-end;
-	overflow: hidden;
-}
-
-.slider_growth {
-	width: 100%;
-	height: calc(100% / 396 * (300 + 3.2 * 0));
-	background: #ffffff;
-	padding: 12px 16px 11px 16px;
-	display: flex;
-	flex-direction: column;
-	justify-content: space-between;
-}
-
-.slider_knob {
-	width: 96px;
-	height: 6px;
-	background: #0e1518;
-	opacity: 0.2;
-	border-radius: 4px;
-	margin: 0 auto;
-	cursor: ns-resize;
-}
-
-.slider_value {
-	height: 24px;
-	font-family: "Inter";
-	font-style: normal;
-	font-weight: 700;
-	font-size: 20px;
-	line-height: 24px;
-	letter-spacing: 0.02em;
-	color: #000000;
-}
-
 .flex_options {
 	height: 100%;
 	display: flex;
@@ -982,34 +489,7 @@ a {
 	gap: 8px;
 }
 
-.skroll_list {
-	position: relative;
-	padding-left: 16px;
-	height: 18px;
-	margin: 34px 0;
-	cursor: pointer;
-}
-
-.flex_gap-8 {
-	display: flex;
-	flex-direction: column;
-	gap: 8px;
-}
-
-.skroll_list_closed::after {
-	position: absolute;
-	content: url(./assets/img/icon/arrow_down_small.svg);
-	top: 0;
-	right: 16px;
-}
-
-.skroll_list_open::after {
-	position: absolute;
-	content: url(./assets/img/icon/arrow_top_small.svg);
-	top: 0;
-	right: 16px;
-}
-
+/* GLOBAL */
 .scroll-fade-enter-active {
 	transition: all 0.8s ease-out;
 }
@@ -1023,17 +503,7 @@ a {
 	transform: translateY(-20px);
 	opacity: 0;
 }
+/* GLOBAL */
 
-/* -------------slider-grwwth------------------- */
-
-.wrapp_wel {
-	max-width: 394px;
-	height: 100%;
-	position: absolute;
-}
-
-.stripe {
-	width: 2px;
-	background-color: rgba(255, 255, 255, 0.1);
-}
+/* ---------------------sidebar_right----------------------*/
 </style>
