+++ src/assets/catalog/texts/dic.js
@@ -27,5 +27,5 @@ export default {
         },
     ],
 
-    select_lang: "ua",
+    select_lang: "ru",
 };
